requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
exceptiongroup==1.1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonpath==0.82
Markdown==3.4.3
model-mommy==2.0.0
packaging==23.1
pluggy==1.2.0
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==2.0.3
